fix tests; use integration for API tests
