+++ firmware.ini
@@ -1,7 +1,7 @@
 [FIRMWARE]
-firmware_version=8ef34717
-firmware_date=2023-03-21
-firmware_sha=8ef34717bfe5e66054f55d88c2fdfe71de3a6d75
+firmware_version=270f680a
+firmware_date=2023-03-23
+firmware_sha=270f680a428372f92e716ec17c4e918fea6974c4
 proto_version=45da982d
 proto_date=2023-03-10
 system_version=3.2.0
